Fixing a gcc build error with std::to_string overloads in JaliStateVector.h

+++ src/statemanager/JaliStateVector.h
@@ -56,7 +56,7 @@ class BaseStateVector {
   //! (which should be forbidden in user-defined names) to avoid
   //! potential collisions with string names if both are used.
   static std::string int_to_string(int const identifier) {
-    return ("~" + std::to_string(identifier)); 
+    return ("~" + std::to_string((int)identifier)); 
   }
 
   //! Query Metadata
